+++ librubyfmt/src/render_targets.rs
@@ -3,11 +3,63 @@ use crate::line_tokens::LineToken;
 use crate::types::{ColNumber, LineNumber};
 use std::collections::HashSet;
 
+fn insert_at<T>(idx: usize, target: &mut Vec<T>, input: &mut Vec<T>) {
+    let drain = input.drain(..);
+    let mut idx = idx;
+    for item in drain {
+        target.insert(idx, item);
+        idx += 1;
+    }
+}
+
+
 #[derive(Copy, Clone, Debug)]
 pub enum ConvertType {
     MultiLine,
     SingleLine,
 }
+
+pub trait LineTokenTarget {
+    fn push(&mut self, lt: LineToken);
+    fn insert_at(&mut self, idx: usize, tokens: &mut Vec<LineToken>);
+    fn into_tokens(self, ct: ConvertType) -> Vec<LineToken>;
+    fn last_token_is_a_newline(&self) -> bool;
+    fn index_of_prev_hard_newline(&self) -> Option<usize>;
+}
+
+#[derive(Debug, Default, Clone)]
+pub struct BaseQueue {
+    tokens: Vec<LineToken>
+}
+
+impl LineTokenTarget for BaseQueue {
+    fn push(&mut self, lt: LineToken) {
+        self.tokens.push(lt)
+    }
+
+    fn insert_at(&mut self, idx: usize, tokens: &mut Vec<LineToken>) {
+        insert_at(idx, &mut self.tokens, tokens)
+    }
+
+    fn into_tokens(self, _ct: ConvertType) -> Vec<LineToken> {
+        self.tokens
+    }
+
+    fn last_token_is_a_newline(&self) -> bool {
+        self.tokens
+            .last()
+            .map(|x| x.is_newline())
+            .unwrap_or(false)
+    }
+
+    fn index_of_prev_hard_newline(&self) -> Option<usize> {
+        self.tokens
+            .iter()
+            .rposition(|v| v.is_newline() || v.is_comment())
+    }
+}
+
+
 
 #[derive(Debug, Clone)]
 pub struct BreakableEntry {
@@ -17,21 +69,16 @@ pub struct BreakableEntry {
     delims: BreakableDelims,
 }
 
-impl BreakableEntry {
-    pub fn new(spaces: ColNumber, delims: BreakableDelims) -> Self {
-        BreakableEntry {
-            spaces,
-            tokens: vec![],
-            line_numbers: HashSet::new(),
-            delims,
-        }
-    }
-
-    pub fn push(&mut self, lt: LineToken) {
+impl LineTokenTarget for BreakableEntry {
+    fn push(&mut self, lt: LineToken) {
         self.tokens.push(lt);
     }
 
-    pub fn into_tokens(self, ct: ConvertType) -> Vec<LineToken> {
+    fn insert_at(&mut self, idx: usize, tokens: &mut Vec<LineToken>) {
+        insert_at(idx, &mut self.tokens, tokens)
+    }
+
+    fn into_tokens(self, ct: ConvertType) -> Vec<LineToken> {
         let mut tokens = self.tokens;
         match ct {
             ConvertType::MultiLine => {
@@ -48,6 +95,30 @@ impl BreakableEntry {
         tokens
     }
 
+    fn last_token_is_a_newline(&self) -> bool {
+        match self.tokens.last() {
+            Some(x) => x.is_newline(),
+            _ => false,
+        }
+    }
+
+    fn index_of_prev_hard_newline(&self) -> Option<usize> {
+        self.tokens
+            .iter()
+            .rposition(|v| v.is_newline() || v.is_comment())
+    }
+}
+
+impl BreakableEntry {
+    pub fn new(spaces: ColNumber, delims: BreakableDelims) -> Self {
+        BreakableEntry {
+            spaces,
+            tokens: vec![],
+            line_numbers: HashSet::new(),
+            delims,
+        }
+    }
+
     pub fn single_line_string_length(&self) -> usize {
         self.tokens
             .iter()
@@ -62,12 +133,5 @@ impl BreakableEntry {
 
     pub fn is_multiline(&self) -> bool {
         self.line_numbers.len() > 1
-    }
-
-    pub fn last_token_is_a_newline(&self) -> bool {
-        match self.tokens.last() {
-            Some(x) => x.is_newline(),
-            _ => false,
-        }
     }
 }
